Use statically linked UpdateElevate

+++ Software/scripts/win32/prepare_installer.sh
@@ -3,11 +3,15 @@
 set -e
 
 # build UpdateElevate
-echo "#define NAME L\"Prismatik\"" > UpdateElevate/UpdateElevate/command.h
-echo "#define EXT L\".exe\"" >> UpdateElevate/UpdateElevate/command.h
-echo "#define ARGS L\" /VERYSILENT /SUPPRESSMSGBOXES /NORESTART\"" >> UpdateElevate/UpdateElevate/command.h
-cmd //c scripts\\win32\\build_UpdateElevate.bat
-cp UpdateElevate/x64/Release/UpdateElevate.exe dist_windows/content/
+if [ -f dist_windows/content/UpdateElevate.exe ]; then
+	echo "Using existing (frozen) UpdateElevate";
+else
+	echo "#define NAME L\"Prismatik\"" > UpdateElevate/UpdateElevate/command.h
+	echo "#define EXT L\".exe\"" >> UpdateElevate/UpdateElevate/command.h
+	echo "#define ARGS L\" /VERYSILENT /SUPPRESSMSGBOXES /NORESTART\"" >> UpdateElevate/UpdateElevate/command.h
+	cmd //c scripts\\win32\\build_UpdateElevate.bat
+	cp UpdateElevate/x64/Release/UpdateElevate.exe dist_windows/content/
+fi
